updated scripts
git-svn-id: svn://svnext.vito.be/ENE/TEM/EMIS/GORD-GSAS/trunk/src@1421 cf4ef618-bcd7-8443-a187-13cdf260f84e

+++ db_reset_gord.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
+# Initialize rbenv and change to this scripts working directory, so rbenv picks up the .ruby-version
 source /etc/profile.d/rbenv.sh
-CUR_DIR="`dirname \"$0\"`"
-cd "$CUR_DIR"
+cd "$(dirname "$0")";
 
-RAILS_ENV=production bundle exec rake db:reset
+# run rake task
+read -p "Are you sure? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    # do dangerous stuff
+    RAILS_ENV=production bundle exec rake db:reset
+fi
